Repaired sending of short SMS messages. Attempted to enable sending of long SMS messages using two separate texts, but only the shorter second text is sending.

+++ Phase/Phase.h
@@ -33,7 +33,7 @@ private:
 	// double delay; // delay between voltage and current
 	signed long delay; // temporary; will be double eventually
 	signed long delaysum;
-	unsigned int numsamples;
+	unsigned long numsamples;
 };
 
 #endif
